# contracts/openapi/lean-gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Obligation Runtime Lean Gateway",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/reviews/{thread_id}": {
      "get": {
        "summary": "Get Review",
        "description": "Return review payload for a thread; 404 if not pending or unknown.",
        "operationId": "get_review_v1_reviews__thread_id__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reviews": {
      "post": {
        "summary": "Create Pending Review",
        "description": "Create or replace pending review (called by orchestrator on interrupt).",
        "operationId": "create_pending_review_v1_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePendingReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePendingReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/reviews/{thread_id}/resume": {
      "post": {
        "summary": "Resume Review",
        "description": "Resume the graph after approve/reject by calling the orchestrator service.",
        "operationId": "resume_review_v1_reviews__thread_id__resume_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResumeProxyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcceptanceSummary": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "0.1.0"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "trust_level": {
            "type": "string",
            "title": "Trust Level",
            "default": "clean"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "build": {
            "type": "object",
            "title": "Build"
          },
          "axiom_audit": {
            "type": "object",
            "title": "Axiom Audit"
          },
          "fresh_checker": {
            "type": "object",
            "title": "Fresh Checker"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AcceptanceSummary",
        "description": "Typed summary of batch (acceptance lane) verification for WitnessBundle.acceptance."
      },
...
      "AxiomAuditResult": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "0.1.0"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "trust_level": {
            "type": "string",
            "enum": [
              "clean",
              "warning",
              "blocked"
            ],
            "title": "Trust Level"
          },
          "blocked_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocked Reasons"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/AxiomDependency"
            },
            "type": "array",
            "title": "Dependencies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok",
          "trust_level"
        ],
        "title": "AxiomAuditResult"
      },
      "AxiomDependency": {
        "properties": {
          "declaration": {
            "type": "string",
            "title": "Declaration"
          },
          "axioms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Axioms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "declaration"
        ],
        "title": "AxiomDependency"
      },
...
      "CreatePendingReviewRequest": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "thread_id"
        ],
        "title": "CreatePendingReviewRequest",
        "description": "Minimum fields to create a pending review; additional keys stored as-is."
      },
      "CreatePendingReviewResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thread_id"
        ],
        "title": "CreatePendingReviewResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyAuditTrail": {
        "properties": {
          "policy_pack_name": {
            "type": "string",
            "title": "Policy Pack Name",
            "default": ""
          },
          "policy_pack_version": {
            "type": "string",
            "title": "Policy Pack Version",
            "default": ""
          },
          "resolved_rules": {
            "items": {
              "$ref": "#/components/schemas/PolicyResolvedRule"
            },
            "type": "array",
            "title": "Resolved Rules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PolicyAuditTrail",
        "description": "Structured policy provenance for review UI and witnesses."
      },
      "PolicyDecision": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "accepted",
              "rejected",
              "blocked",
              "needs_review",
              "lower_trust",
              "failed"
            ],
            "title": "Decision"
          },
          "trust_level": {
            "type": "string",
            "enum": [
              "clean",
              "warning",
              "blocked"
            ],
            "title": "Trust Level"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "policy_pack_name": {
            "type": "string",
            "title": "Policy Pack Name",
            "default": ""
          },
          "policy_pack_version": {
            "type": "string",
            "title": "Policy Pack Version",
            "default": ""
          },
          "resolved_rules": {
            "items": {
              "$ref": "#/components/schemas/PolicyResolvedRule"
            },
            "type": "array",
            "title": "Resolved Rules"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "trust_level"
        ],
        "title": "PolicyDecision"
      },
      "PolicyResolvedRule": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "source_pack": {
            "type": "string",
            "title": "Source Pack"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched"
          },
          "effect": {
            "type": "string",
            "enum": [
              "none",
              "needs_review",
              "rejected",
              "blocked",
              "accepted"
            ],
            "title": "Effect"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule_id",
          "source_pack",
          "matched",
          "effect",
          "reason_code"
        ],
        "title": "PolicyResolvedRule",
        "description": "One evaluated policy rule for audit trails."
      },
...
      "ReviewEnvironmentSummary": {
        "properties": {
          "repo_id": {
            "type": "string",
            "title": "Repo Id",
            "default": ""
          },
          "repo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Url"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha",
            "default": ""
          },
          "lean_toolchain": {
            "type": "string",
            "title": "Lean Toolchain",
            "default": ""
          },
          "lakefile_hash": {
            "type": "string",
            "title": "Lakefile Hash",
            "default": ""
          },
          "manifest_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manifest Hash"
          },
          "target_platform": {
            "type": "string",
            "title": "Target Platform",
            "default": "linux/amd64"
          },
          "build_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Build Flags"
          },
          "os_family": {
            "type": "string",
            "title": "Os Family",
            "default": "linux"
          },
          "imported_packages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imported Packages"
          }
        },
        "type": "object",
        "title": "ReviewEnvironmentSummary",
        "description": "Environment fingerprint subset for review display."
      },
      "ReviewObligationSummary": {
        "properties": {
          "obligation": {
            "type": "object",
            "title": "Obligation"
          },
          "target_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Files"
          },
          "target_declarations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Declarations"
          }
        },
        "type": "object",
        "title": "ReviewObligationSummary",
        "description": "Obligation context shown in the review UI."
      },
      "ReviewPatchMetadata": {
        "properties": {
          "current_patch": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Current Patch"
          },
          "protected_paths_touched": {
            "type": "boolean",
            "title": "Protected Paths Touched",
            "default": false
          },
          "imports_changed": {
            "type": "boolean",
            "title": "Imports Changed",
            "default": false
          },
          "changed_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed Files"
          },
          "diff_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff Hash"
          }
        },
        "type": "object",
        "title": "ReviewPatchMetadata",
        "description": "Patch and diff context for reviewers."
      },
      "ReviewPayload": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "obligation_id": {
            "type": "string",
            "title": "Obligation Id",
            "default": ""
          },
          "obligation_summary": {
            "$ref": "#/components/schemas/ReviewObligationSummary"
          },
          "environment_summary": {
            "$ref": "#/components/schemas/ReviewEnvironmentSummary"
          },
          "patch_metadata": {
            "$ref": "#/components/schemas/ReviewPatchMetadata"
          },
          "diff_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff Summary"
          },
          "diagnostics_summary": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Diagnostics Summary"
          },
          "axiom_audit_summary": {
            "$ref": "#/components/schemas/AxiomAuditResult"
          },
          "batch_summary": {
            "$ref": "#/components/schemas/AcceptanceSummary"
          },
          "policy_summary": {
            "$ref": "#/components/schemas/PolicyDecision"
          },
          "trust_delta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Delta"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "awaiting_review"
          },
          "policy_audit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyAuditTrail"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "thread_id"
        ],
        "title": "ReviewPayload",
        "description": "Payload served to the review UI for interrupted obligations."
      },
      "ReviewResumeProxyResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "status": {
            "title": "Status"
          },
          "artifacts_count": {
            "type": "integer",
            "title": "Artifacts Count",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thread_id"
        ],
        "title": "ReviewResumeProxyResponse",
        "description": "Shape returned by orchestrator resume; gateway proxies this."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}